+++ components/js/base/base.js
@@ -1,0 +1,13 @@
+this.Spellbook.Classes.Base = (function() {
+  Base.prototype._settings = {};
+
+  function Base(options) {
+    this.options = options;
+    if (typeof this.init === "function") {
+      this.init();
+    }
+  }
+
+  return Base;
+
+})();