+++ components/js/classes/modal.js
@@ -1,15 +1,18 @@
-this.Spellbook.Classes.Modal = (function() {
+var extend = function(child, parent) { for (var key in parent) { if (hasProp.call(parent, key)) child[key] = parent[key]; } function ctor() { this.constructor = child; } ctor.prototype = parent.prototype; child.prototype = new ctor(); child.__super__ = parent.prototype; return child; },
+  hasProp = {}.hasOwnProperty;
+
+this.Spellbook.Classes.Modal = (function(superClass) {
+  extend(Modal, superClass);
+
+  function Modal() {
+    return Modal.__super__.constructor.apply(this, arguments);
+  }
+
   Modal.prototype._$modal = null;
 
   Modal.prototype._$backdrop = null;
 
-  Modal.prototype._settings = {};
-
-  function Modal(options) {
-    this.init(options);
-  }
-
-  Modal.prototype.init = function(options) {
+  Modal.prototype.init = function() {
     this._settings = $.extend({
       $trigger: $('.js-modal-trigger'),
       $close: $('.js-modal-close'),
@@ -18,7 +21,7 @@ this.Spellbook.Classes.Modal = (function() {
       activeClass: 'is-active',
       inactiveClass: 'is-inactive',
       activeBodyClass: 'is-modal-active'
-    }, options);
+    }, this.options);
     return this._setEventHandlers();
   };
 
@@ -110,4 +113,4 @@ this.Spellbook.Classes.Modal = (function() {
 
   return Modal;
 
-})();
+})(Spellbook.Classes.Base);