+++ components/js/classes/quantityInput.js
@@ -1,13 +1,16 @@
-this.Spellbook.Classes.QuantityInput = (function() {
-  QuantityInput.prototype._settings = {};
+var extend = function(child, parent) { for (var key in parent) { if (hasProp.call(parent, key)) child[key] = parent[key]; } function ctor() { this.constructor = child; } ctor.prototype = parent.prototype; child.prototype = new ctor(); child.__super__ = parent.prototype; return child; },
+  hasProp = {}.hasOwnProperty;
+
+this.Spellbook.Classes.QuantityInput = (function(superClass) {
+  extend(QuantityInput, superClass);
+
+  function QuantityInput() {
+    return QuantityInput.__super__.constructor.apply(this, arguments);
+  }
 
   QuantityInput.prototype._value = null;
 
-  function QuantityInput(options) {
-    this.init(options);
-  }
-
-  QuantityInput.prototype.init = function(options) {
+  QuantityInput.prototype.init = function() {
     this._settings = $.extend({
       $element: $('.js-quantityInput'),
       $field: $('.js-quantityInput-field'),
@@ -21,7 +24,7 @@ this.Spellbook.Classes.QuantityInput = (function() {
       onIncrease: null,
       onDecrease: null,
       onTargetUpdate: null
-    }, options);
+    }, this.options);
     this._setValue();
     return this._setEventHandlers();
   };
@@ -88,4 +91,4 @@ this.Spellbook.Classes.QuantityInput = (function() {
 
   return QuantityInput;
 
-})();
+})(Spellbook.Classes.Base);