+++ components/js/classes/selectText.js
@@ -1,15 +1,18 @@
-this.Spellbook.Classes.SelectText = (function() {
-  SelectText.prototype._settings = {};
+var extend = function(child, parent) { for (var key in parent) { if (hasProp.call(parent, key)) child[key] = parent[key]; } function ctor() { this.constructor = child; } ctor.prototype = parent.prototype; child.prototype = new ctor(); child.__super__ = parent.prototype; return child; },
+  hasProp = {}.hasOwnProperty;
 
-  function SelectText(options) {
-    this.init(options);
+this.Spellbook.Classes.SelectText = (function(superClass) {
+  extend(SelectText, superClass);
+
+  function SelectText() {
+    return SelectText.__super__.constructor.apply(this, arguments);
   }
 
-  SelectText.prototype.init = function(options) {
+  SelectText.prototype.init = function() {
     this._settings = $.extend({
       $element: $('.js-selectText'),
       onClick: null
-    }, options);
+    }, this.options);
     return this._setEventHandlers();
   };
 
@@ -42,4 +45,4 @@ this.Spellbook.Classes.SelectText = (function() {
 
   return SelectText;
 
-})();
+})(Spellbook.Classes.Base);